Add XML comments and decrease namespace indentation
