+++ Enumeration.EFCore.Tests/DbContext/TestDbContext.cs
@@ -1,28 +1,27 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using PM.Enumeration.EFCore;
 
-namespace Enumeration.EFCore.Tests.DbContext
+namespace Enumeration.EFCore.Tests.DbContext;
+
+public class TestDbContext : Microsoft.EntityFrameworkCore.DbContext
 {
-    public class TestDbContext : Microsoft.EntityFrameworkCore.DbContext
+    public DbSet<TestEntity> TestEntities { get; set; } = null!;
+
+    public TestDbContext(DbContextOptions<TestDbContext> options) : base(options)
     {
-        public DbSet<TestEntity> TestEntities { get; set; } = null!;
+    }
 
-        public TestDbContext(DbContextOptions<TestDbContext> options) : base(options)
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        base.OnModelCreating(modelBuilder);
+
+        modelBuilder.Entity<TestEntity>(e =>
         {
-        }
+            e.Property(p => p.Test)
+                .HasConversion(new EnumerationConverter<TestEnumeration>());
 
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            base.OnModelCreating(modelBuilder);
-
-            modelBuilder.Entity<TestEntity>(e =>
-            {
-                e.Property(p => p.Test)
-                    .HasConversion(new EnumerationConverter<TestEnumeration>());
-                
-                e.Property(p => p.TestDynamic)
-                    .HasConversion(new EnumerationDynamicConverter<TestEnumerationDynamic>());
-            });
-        }
+            e.Property(p => p.TestDynamic)
+                .HasConversion(new EnumerationDynamicConverter<TestEnumerationDynamic>());
+        });
     }
 }
