+++ Enumeration.EFCore.Tests/DbContext/TestEntity.cs
@@ -1,36 +1,35 @@
 ﻿using System;
 using PM.Enumeration;
 
-namespace Enumeration.EFCore.Tests.DbContext
+namespace Enumeration.EFCore.Tests.DbContext;
+
+public class TestEntity
 {
-    public class TestEntity
+    public Guid Id { get; set; }
+
+    public TestEnumeration? Test { get; set; }
+
+    public TestEnumerationDynamic? TestDynamic { get; set; }
+}
+
+public class TestEnumeration : Enumeration<TestEnumeration>
+{
+    public static readonly TestEnumeration CodeA = new(nameof(CodeA));
+
+    private TestEnumeration(string value) : base(value)
     {
-        public Guid Id { get; set; }
-        
-        public TestEnumeration? Test { get; set; }
-        
-        public TestEnumerationDynamic? TestDynamic { get; set; }
+    }
+}
+
+public class TestEnumerationDynamic : EnumerationDynamic<TestEnumerationDynamic>
+{
+    public static readonly TestEnumerationDynamic CodeA = new(nameof(CodeA));
+
+    public TestEnumerationDynamic()
+    {
     }
 
-    public class TestEnumeration : Enumeration<TestEnumeration>
+    private TestEnumerationDynamic(string value) : base(value)
     {
-        public static readonly TestEnumeration CodeA = new(nameof(CodeA));
-        
-        private TestEnumeration(string value) : base(value)
-        {
-        }
-    }
-    
-    public class TestEnumerationDynamic : EnumerationDynamic<TestEnumerationDynamic>
-    {
-        public static readonly TestEnumerationDynamic CodeA = new(nameof(CodeA));
-
-        public TestEnumerationDynamic()
-        {
-        }
-
-        private TestEnumerationDynamic(string value) : base(value)
-        {
-        }
     }
 }
