+++ Enumeration.EFCore.Tests/Enumeration.EFCore.Tests.csproj
@@ -22,10 +22,6 @@
     </ItemGroup>
 
     <ItemGroup>
-      <Compile Remove="UnitTest1.cs" />
-    </ItemGroup>
-
-    <ItemGroup>
       <ProjectReference Include="..\Enumeration.EFCore\Enumeration.EFCore.csproj" />
     </ItemGroup>
