+++ Enumeration.sln
@@ -12,6 +12,16 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Enumeration.EFCore.Tests", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Enumeration.SystemTextJson.Tests", "Enumeration.SystemTextJson.Tests\Enumeration.SystemTextJson.Tests.csproj", "{42321360-16E0-4B60-B52F-6F86C9153E95}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "1. Src", "1. Src", "{D5F0A5F7-1947-4E19-9946-436CEA9A5D95}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "2. Tests", "2. Tests", "{6FDC7463-F8B8-4EDF-8FC5-78D544DD83AC}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "3. Docs", "3. Docs", "{29E02E5C-3AE5-4EFF-9E61-42169F7F174B}"
+	ProjectSection(SolutionItems) = preProject
+		LICENSE = LICENSE
+		README.md = README.md
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -42,5 +52,13 @@ Global
 		{42321360-16E0-4B60-B52F-6F86C9153E95}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{42321360-16E0-4B60-B52F-6F86C9153E95}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{42321360-16E0-4B60-B52F-6F86C9153E95}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{4BD298FF-EB3E-4295-9B25-11F16DE8CA37} = {D5F0A5F7-1947-4E19-9946-436CEA9A5D95}
+		{BCA6E329-1A9D-49EA-9D3E-21240D4E4C5F} = {D5F0A5F7-1947-4E19-9946-436CEA9A5D95}
+		{7F2BBBBB-1AEB-49D2-826A-A42A8074792B} = {D5F0A5F7-1947-4E19-9946-436CEA9A5D95}
+		{B8255C46-FFC0-42C7-97CE-BED81EBE09C7} = {6FDC7463-F8B8-4EDF-8FC5-78D544DD83AC}
+		{42321360-16E0-4B60-B52F-6F86C9153E95} = {6FDC7463-F8B8-4EDF-8FC5-78D544DD83AC}
+		{F13894B2-F6AD-4E89-9F2C-72A09053EA00} = {6FDC7463-F8B8-4EDF-8FC5-78D544DD83AC}
 	EndGlobalSection
 EndGlobal
