+++ README.md
@@ -1,9 +1,7 @@
 # Enumeration
 
+##TODO
 - ReadMe
-- Comments
 - Test .net 6
 - Code coverage
-- internals where possible
-- warnings
 - newtonsoft converters and tests
